restructure and modularize for better readability
